Add generic patchIndigoUser utility function
